// src/gen/api/types/CreatePostBody.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export const createPostBodyVisibilityEnum = {
  public: 'public',
  private: 'private',
} as const

export type CreatePostBodyVisibilityEnumKey =
  (typeof createPostBodyVisibilityEnum)[keyof typeof createPostBodyVisibilityEnum]

export type CreatePostBody = {
  /**
   * @maxLength 10000
   * @type string
   */
  content: string
  /**
   * @type string
   */
  visibility: CreatePostBodyVisibilityEnumKey
  /**
   * @type array | undefined
   */
  images?: {
    /**
     * @minLength 1
     * @type string
     */
    r2Key: string
    /**
     * @type integer
     */
    width: number
    /**
     * @type integer
     */
    height: number
    /**
     * @minLength 0
     * @maxLength 19
     * @type integer
     */
    position: number
  }[]
}