// src/gen/api/types/PresignBody.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export const presignBodyContentTypesEnum = {
  'image/jpeg': 'image/jpeg',
  'image/png': 'image/png',
  'image/webp': 'image/webp',
} as const

export type PresignBodyContentTypesEnumKey =
  (typeof presignBodyContentTypesEnum)[keyof typeof presignBodyContentTypesEnum]

export type PresignBody = {
  /**
   * @minLength 1
   * @maxLength 20
   * @type integer
   */
  count: number
  /**
   * @type array
   */
  contentTypes: PresignBodyContentTypesEnumKey[]
}